Aggiunto check del form crea ordine, controlli effettutati sia lato client sia lato server. Quando si verifica un errore viene mostrato in errorpage. Da sistemare: css Errorepage perchè gli errori si accavallano

+++ src/main/java/control/CreaOrdine.java
@@ -37,50 +37,8 @@ public class CreaOrdine extends HttpServlet {
         int totale=(int)request.getSession().getAttribute("totale");
 
         List<String> errori=new ArrayList<>();
-        //controlli
-        if(via.isEmpty()){
-            errori.add("campila il campo indirizzo");
-        }
-        if(city.isEmpty()){
-            errori.add("compila il campo città");
-        }
-        if(cap.isEmpty()){
-            errori.add("Compila il campo cap");
-        }
-        if(card.isEmpty()){
-            errori.add("compila il campo numero carta");
-        }
-        if(scad.isEmpty()){
-            errori.add("compila il campo scadenza");
-        }
-        if(cvc.isEmpty()){
-            errori.add("compila il campo cvc");
-        }
-        if(!isValidAddress(via)){
-            errori.add("compila correttamente il campo indirizzo");
-        }
-        if(!isValidCity(city)){
-            errori.add("compila correttamente il campo citta");
-        }
-        if(!isValidCap(cap)){
-            errori.add("compila correttamente il campo cap");
-        }
-        if(!isValidCard(card)){
-            errori.add("compila correttamente il campo numero carta");
-        }
-        if(!isValidScad(scad)){
-            errori.add("compila correttamente il campo scadenza");
-        }
-        if(!isValidCvc(cvc)){
-            errori.add("compila correttamente il campo cvc");
-        }
 
-        if(!errori.isEmpty()){
-            request.getSession().setAttribute("errori",errori);
-            address="WEB-INF/Errorepage.jsp";
-            RequestDispatcher rd= request.getRequestDispatcher(address);
-            rd.forward(request,response);
-        }else {
+
 
 
             Ordine ordine = new Ordine();
@@ -95,26 +53,72 @@ public class CreaOrdine extends HttpServlet {
             String prodotti = "";
 
             if (button.equals("back")) {
-                address = "/WEB-INF/carrello.jsp";
+                RequestDispatcher rd = request.getRequestDispatcher("/WEB-INF/carrello.jsp");
+                rd.forward(request, response);
             } else {
-                ordine.setIdCliente(user.getUsername());
-                ordine.setIdCarrello(carrello.getIdCarrello());
-                ordine.setSpedizione(via + " " + city + " " + cap);
-                ordine.setPagamento("Numero Carta: " + card + " Scad: " + scad + " CVC: " + cvc);
-                ordine.setTotale(totale);
-                ordine.setDataOrd(data);
-                pc = pcdao.doRetriveAllById(carrello.getIdCarrello());
-                for (ProdottiCarrello ps : pc) {
-                    prodotti += "Id prodotto: " + Integer.toString(ps.getIdProdotto()) + ", quantita': " + Integer.toString(ps.getQuantita()) + ", taglia: " + ps.getTaglia() + "\r\n";
+                //controlli
+                if(via.isEmpty()){
+                    errori.add("campila il campo indirizzo");
                 }
-                ordine.setProdotti(prodotti);
-                ordineDAO.doSave(ordine);
-                address = "/WEB-INF/index.jsp";
-                pdao.deleteProdottoCarrelloByCarrello(carrello.getIdCarrello());
+                if(city.isEmpty()){
+                    errori.add("compila il campo città");
+                }
+                if(cap.isEmpty()){
+                    errori.add("Compila il campo cap");
+                }
+                if(card.isEmpty()){
+                    errori.add("compila il campo numero carta");
+                }
+                if(scad.isEmpty()){
+                    errori.add("compila il campo scadenza");
+                }
+                if(cvc.isEmpty()){
+                    errori.add("compila il campo cvc");
+                }
+                if(!isValidAddress(via)){
+                    errori.add("compila correttamente il campo indirizzo");
+                }
+                if(!isValidCity(city)){
+                    errori.add("compila correttamente il campo citta");
+                }
+                if(!isValidCap(cap)){
+                    errori.add("compila correttamente il campo cap");
+                }
+                if(!isValidCard(card)){
+                    errori.add("compila correttamente il campo numero carta");
+                }
+                if(!isValidScad(scad)){
+                    errori.add("compila correttamente il campo scadenza");
+                }
+                if(!isValidCvc(cvc)){
+                    errori.add("compila correttamente il campo cvc");
+                }
+
+                if(!errori.isEmpty()){
+                    request.getSession().setAttribute("errori",errori);
+                    RequestDispatcher rd= request.getRequestDispatcher("WEB-INF/Errorepage.jsp");
+                    rd.forward(request,response);
+                }else {
+                    ordine.setIdCliente(user.getUsername());
+                    ordine.setIdCarrello(carrello.getIdCarrello());
+                    ordine.setSpedizione(via + " " + city + " " + cap);
+                    ordine.setPagamento("Numero Carta: " + card + " Scad: " + scad + " CVC: " + cvc);
+                    ordine.setTotale(totale);
+                    ordine.setDataOrd(data);
+                    pc = pcdao.doRetriveAllById(carrello.getIdCarrello());
+                    for (ProdottiCarrello ps : pc) {
+                        prodotti += "Id prodotto: " + Integer.toString(ps.getIdProdotto()) + ", quantita': " + Integer.toString(ps.getQuantita()) + ", taglia: " + ps.getTaglia() + "\r\n";
+                    }
+                    ordine.setProdotti(prodotti);
+                    ordineDAO.doSave(ordine);
+                    address = "/WEB-INF/index.jsp";
+                    pdao.deleteProdottoCarrelloByCarrello(carrello.getIdCarrello());
+                    RequestDispatcher rd = request.getRequestDispatcher(address);
+                    rd.forward(request, response);
+                }
             }
-        }
-        RequestDispatcher rd = request.getRequestDispatcher(address);
-        rd.forward(request, response);
+
+
     }
     public static boolean isValidAddress(String password) {
         Pattern pattern = Pattern.compile("^[a-zA-Z ]{3,}$");
@@ -125,7 +129,7 @@ public class CreaOrdine extends HttpServlet {
         return pattern.matcher(password).matches();
     }
     public static boolean isValidCap(String password) {
-        Pattern pattern = Pattern.compile("^[0-9]{3}$");
+        Pattern pattern = Pattern.compile("^[0-9]{3,}$");
         return pattern.matcher(password).matches();
     }
     public static boolean isValidCard(String password) {
